Melhora na logica do cupom, imagem de gabinete e ajustes gerais

+++ src/main/java/dados/IRepositorioPedido.java
@@ -11,7 +11,7 @@ public interface IRepositorioPedido {
     void salvar(Pedido pedido);
     void atualizar(Pedido pedido);
     void deletar(int id) throws PedidoNaoEncontradoException;
-    Pedido buscarPorId(int id) throws PedidoNaoEncontradoException;
+    Pedido buscarPorId(long id) throws PedidoNaoEncontradoException;
     List<Pedido> listarTodos();
     List<Pedido> buscarPorCliente(Cliente cliente);
